Split evidence_calculator.py into likelihood_calculator.py and sample_processing.py
- Created likelihood_calculator.py with LikelihoodCalculator class and mathematical utilities
- Created sample_processing.py with IntrinsicSampleProcessor and ExtrinsicSampleProcessor classes
- Updated all import statements across the codebase to use new module structure
- Improved separation of concerns: likelihood calculations vs sample handling
- No functionality lost, all code preserved in appropriate modules
